requirements.txt
Keras==2.15.0
Pillow==10.1.0
numpy
scikit-learn
tensorflow
pandas
matplotlib
kaggle
seaborn